moved class-based tests to "tests" folder
