--- integration_tests/tests2/test_emr_serverless.py
@@ -1,11 +1,0 @@
-from lib.core.constants import SettingConfigResourceTypes
-
-from test_base_class import TestBaseClass
-
-
-class TestEMRServerless(TestBaseClass):
-    __test__ = True  # to override BaseClass' skip
-
-    @classmethod
-    def setup_class(cls):
-        cls.resource_type = SettingConfigResourceTypes.EMR_SERVERLESS
